+++ src/main/java/roomescape/reservation/domain/aggregate/ReservationName.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.domain;
+package roomescape.reservation.domain.aggregate;
 
 public final class ReservationName {
     private final String name;
