+++ src/main/java/roomescape/reservation/infrastructure/dao/ReservationTimeDao.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.infrastructure;
+package roomescape.reservation.infrastructure.dao;
 
 import java.sql.PreparedStatement;
 import java.time.LocalTime;
@@ -9,8 +9,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.reservation.application.ReservationTimeRepository;
-import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.application.repository.ReservationTimeRepository;
+import roomescape.reservation.domain.aggregate.ReservationTime;
 
 @Repository
 public class ReservationTimeDao implements ReservationTimeRepository {
