+++ src/test/java/roomescape/reservation/infrastructure/ReservationDaoTest.java
@@ -12,9 +12,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.reservation.application.dto.CreateReservationRequest;
-import roomescape.reservation.domain.ReservationDate;
-import roomescape.reservation.domain.ReservationName;
-import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.domain.aggregate.ReservationDate;
+import roomescape.reservation.domain.aggregate.ReservationName;
+import roomescape.reservation.infrastructure.dao.ReservationDao;
+import roomescape.reservation.infrastructure.dao.ReservationTimeDao;
 import roomescape.reservation.presentation.dto.ReservationTimeRequest;
 
 @JdbcTest
