+++ src/test/java/roomescape/reservation/infrastructure/ReservationTimeDaoTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.infrastructure.dao.ReservationTimeDao;
 import roomescape.reservation.presentation.dto.ReservationTimeRequest;
 
 @JdbcTest
